requirements.txt
asgiref==3.5.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.2
decorator==5.1.1
Deprecated==1.2.13
Django==4.0.4
django-ckeditor==6.4.0
django-cors-headers==3.12.0
django-debug-toolbar==3.2.4
django-js-asset==2.0.0
django-oauth-toolkit==1.7.0
djangorestframework==3.13.1
drf-yasg==1.20.0
generics==2.0.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
jwcrypto==1.0
MarkupSafe==2.1.0
mysqlclient==2.1.0
oauthlib==3.2.0
packaging==21.3
permission==0.4.1
Pillow==9.1.0
pycparser==2.21
PyMySQL==1.0.2
pyparsing==3.0.7
pytz==2022.1
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
self==2020.12.3
sqlparse==0.4.2
tzdata==2022.1
ui==0.1.4
uritemplate==4.1.1
urllib3==1.26.8
wrapt==1.14.0
